Initial refactor existing guide (ssh)
