+++ web/packages/teleport/src/Bots/Add/GitHubActionsSsh/AddBotToWorkflow.story.tsx
@@ -19,52 +19,31 @@
 import { MemoryRouter } from 'react-router';
 
 import { ContextProvider } from 'teleport';
-import { allAccessAcl, createTeleportContext } from 'teleport/mocks/contexts';
+import { createTeleportContext } from 'teleport/mocks/contexts';
 
 import { FlowStepProps } from '../Shared/GuidedFlow';
-import { ConfigureBot } from './ConfigureBot';
-import { GitHubFlowProvider } from './useGitHubFlow';
+import { AddBotToWorkflow } from './AddBotToWorkflow';
+import { GitHubSshFlowProvider } from './useGitHubSshFlow';
 
-const Provider = ({ children, access = allAccessAcl }) => {
+const Provider = ({ children }) => {
   const ctx = createTeleportContext();
-  ctx.storeUser.setState({
-    username: 'joe@example.com',
-    acl: access,
-  });
 
   return (
     <MemoryRouter>
       <ContextProvider ctx={ctx}>
-        <GitHubFlowProvider>{children}</GitHubFlowProvider>
+        <GitHubSshFlowProvider>{children}</GitHubSshFlowProvider>
       </ContextProvider>
     </MemoryRouter>
   );
 };
 
 export default {
-  title: 'Teleport/Bots/Add/GitHubActions/Permissions',
+  title: 'Teleport/Bots/Add/GitHubActions+SSH',
 };
 
-export const Loaded = () => (
+export const AddBot = () => (
   <Provider>
-    <ConfigureBot {...props} />
-  </Provider>
-);
-
-export const NoCreatePermission = () => (
-  <Provider
-    access={{
-      ...allAccessAcl,
-      bots: {
-        list: false,
-        read: false,
-        edit: false,
-        create: false,
-        remove: false,
-      },
-    }}
-  >
-    <ConfigureBot {...props} />
+    <AddBotToWorkflow {...props} />
   </Provider>
 );
