+++ web/packages/teleport/src/Bots/Add/GitHubActionsSsh/AddBotToWorkflow.test.tsx
@@ -21,44 +21,38 @@ import { MemoryRouter } from 'react-router-dom';
 import { render, screen } from 'design/utils/testing';
 
 import { ContextProvider } from 'teleport';
-import cfg from 'teleport/config';
+import { allAccessAcl } from 'teleport/mocks/contexts';
 import TeleportContext from 'teleport/teleportContext';
 
-import { Finish } from './Finish';
-import { GitHubFlowProvider, initialBotState } from './useGitHubFlow';
+import { ConfigureBot } from './ConfigureBot';
+import { GitHubSshFlowProvider } from './useGitHubSshFlow';
 
-describe('finish Component', () => {
-  const setup = ({ botName }) => {
+describe('addBotToWorkflow Component', () => {
+  const setup = () => {
     const ctx = new TeleportContext();
+
+    ctx.storeUser.setState({
+      username: 'joe@example.com',
+      acl: allAccessAcl,
+    });
     render(
       <MemoryRouter>
         <ContextProvider ctx={ctx}>
-          <GitHubFlowProvider bot={{ ...initialBotState, botName }}>
-            <Finish />
-          </GitHubFlowProvider>
+          <GitHubSshFlowProvider>
+            <ConfigureBot />
+          </GitHubSshFlowProvider>
         </ContextProvider>
       </MemoryRouter>
     );
   };
 
-  it('renders with dynamic content based on hook', () => {
-    setup({ botName: 'test-bot' });
-    expect(
-      screen.getByText('Your Bot is Added to Teleport')
-    ).toBeInTheDocument();
-    expect(screen.getByText('View Bots')).toBeInTheDocument();
-    expect(screen.getByText('Add Another Bot')).toBeInTheDocument();
+  it('does not display the button to go back', () => {
+    setup();
+    expect(screen.queryByTestId('button-back')).not.toBeInTheDocument();
   });
 
-  it('has correct links on buttons', () => {
-    setup({ botName: 'test-bot' });
-    expect(screen.getByText('View Bots').closest('a')).toHaveAttribute(
-      'href',
-      cfg.getBotsRoute()
-    );
-    expect(screen.getByText('Add Another Bot').closest('a')).toHaveAttribute(
-      'href',
-      cfg.getBotsNewRoute()
-    );
+  it('displays the button to finish', () => {
+    setup();
+    expect(screen.getByTestId('button-next')).toBeInTheDocument();
   });
 });
