+++ web/packages/teleport/src/Bots/Add/GitHubActionsSsh/index.ts
@@ -16,5 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { GitHubActions as default } from './GitHubActions';
+export { GitHubActionsSsh as default } from './GitHubActionsSsh';
 export { ConnectGitHub as ConfigureBot } from './ConnectGitHub';
